Add C++ solution for Missing_Number

+++ Introductory_Problems/Missing_Number/Missing_Number.cpp
@@ -12,9 +12,9 @@ using namespace std;
 int main(){
     int n; 
     cin>>n; 
-    long long sum = n*(n+1)/2; 
+    long long sum = (long long)n*(n+1)/2; 
     for (int i=0; i<n-1; i++){
-        long long x; 
+        int x; 
         cin>>x; 
         sum -= x; 
     }
